+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Recommendation: for ease of reading, use the following order:
 - Fixed
 -->
 
+## [Unreleased]
+## Changed
+- Moved versioned file logic from GQL level to `use_case`
+
 ## [0.246.0] - 2025-08-20
 ### Added
 - Webhooks are integrated into the Flow system, and support retries:
@@ -146,7 +150,6 @@ Recommendation: for ease of reading, use the following order:
 - Configuration, predefined accounts: `password` field is now mandatory (#1277).
 - `kamu`: a start-up job by `PredefinedAccountsRegistrator` will be called earlier, right after creating
   `base_catalog` (#1277).
-- Refactor: move logic for versioned files from GQL level to use_case
 ### Fixed
 - `CreateAccountUseCaseImpl`: added sending `AccountLifecycleMessage`.
 - Renaming `provider_identity_key` field with account name for password accounts.